fix: 테이블 변경사항 반영(댓글갯수), 이벤트 id 타입수정

+++ server/src/main/java/ynzmz/server/board/event/our/controller/OurController.java
@@ -35,7 +35,7 @@ public class OurController {
     }
 
     @PatchMapping("/{event-id}")
-    public ResponseEntity<?> patchOur(@PathVariable("event-id") int id,@RequestBody OurDto.Patch patch){
+    public ResponseEntity<?> patchOur(@PathVariable("event-id") long id,@RequestBody OurDto.Patch patch){
     Our event= mapper.ourPatchToOur(patch);
     event.setEventId(id);
     service.updateEvent(event);
